# back/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/suggest": {"get": {"summary": "Suggest", "operationId": "suggest_suggest_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/SuggestResult"}, "type": "array", "title": "Response Suggest Suggest Get"}}}}}}}, "/patients": {"get": {"summary": "Get Patients", "description": "Get list of patients", "operationId": "get_patients_patients_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Patient"}, "type": "array", "title": "Response Get Patients Patients Get"}}}}}}}, "/patients/{patient_id}": {"get": {"summary": "Get Patient", "description": "Get detail of a patient", "operationId": "get_patient_patients__patient_id__get", "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"title": "Patient Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Patient"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/patients/{patient_id}/history": {"get": {"summary": "Get Patient History", "description": "Get history of this patient", "operationId": "get_patient_history_patients__patient_id__history_get", "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"title": "Patient Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PatientHistory"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/patients/{patient_id}/futures": {"get": {"summary": "Get Patient Futures", "description": "Get possible future trajectories for this patient", "operationId": "get_patient_futures_patients__patient_id__futures_get", "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Patient Id"}}, {"name": "snapshot_events", "in": "query", "required": false, "schema": {"type": "integer", "title": "Snapshot Events"}}, {"name": "top_k", "in": "query", "required": false, "schema": {"type": "integer", "default": 5, "title": "Top K"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/PatientFuture"}, "title": "Response Get Patient Futures Patients  Patient Id  Futures Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/patients/{patient_id}/ews": {"get": {"summary": "Get Patient Ews", "description": "Get a list of possible DRGs that could happen to this person within given time frame", "operationId": "get_patient_ews_patients__patient_id__ews_get", "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"title": "Patient Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/EWS"}, "title": "Response Get Patient Ews Patients  Patient Id  Ews Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"DRG": {"properties": {"code": {"type": "string", "title": "Code"}, "label": {"type": "string", "title": "Label"}}, "type": "object", "required": ["code", "label"], "title": "DRG"}, "EWS": {"properties": {"drg": {"$ref": "#/components/schemas/DRG"}, "probability": {"type": "number", "title": "Probability"}, "eta": {"type": "integer", "title": "Eta"}}, "type": "object", "required": ["drg", "probability", "eta"], "title": "EWS"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthService": {"properties": {"label": {"type": "string", "title": "Label"}, "type": {"$ref": "#/components/schemas/HealthServiceType"}, "delta_days": {"type": "integer", "title": "Delta Days"}, "detail": {"title": "Detail"}}, "type": "object", "required": ["label", "type", "delta_days", "detail"], "title": "HealthService"}, "HealthServiceType": {"type": "string", "enum": ["PROCEDURE", "MEDICATION", "HEALTH_TOOL", "STOMATOLOGICAL_TOOL", "TRANSPORT", "HOSPITALIZATION", "SPA", "DEATH", "UNKNOWN"], "title": "HealthServiceType"}, "Patient": {"properties": {"patient_id": {"type": "string", "title": "Patient Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["patient_id", "name"], "title": "Patient"}, "PatientFuture": {"properties": {"expected_health_services": {"items": {"$ref": "#/components/schemas/HealthService"}, "type": "array", "title": "Expected Health Services"}, "probability": {"type": "number", "title": "Probability"}}, "type": "object", "required": ["expected_health_services", "probability"], "title": "PatientFuture"}, "PatientHistory": {"properties": {"received_health_services": {"items": {"$ref": "#/components/schemas/HealthService"}, "type": "array", "title": "Received Health Services"}}, "type": "object", "required": ["received_health_services"], "title": "PatientHistory"}, "SuggestResult": {"properties": {"label": {"type": "string", "title": "Label"}, "type": {"$ref": "#/components/schemas/SuggestResultType"}}, "type": "object", "required": ["label", "type"], "title": "SuggestResult"}, "SuggestResultType": {"type": "string", "enum": ["PATIENT", "SERVICE_PROVIDER", "DRG"], "title": "SuggestResultType"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}