# front/src/lib/api-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:8000' | (string & {});
};

/**
 * DRG
 */
export type Drg = {
    /**
     * Code
     */
    code: string;
    /**
     * Label
     */
    label: string;
};

/**
 * EWS
 */
export type Ews = {
    drg: Drg;
    /**
     * Probability
     */
    probability: number;
    /**
     * Eta
     */
    eta: number;
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * HealthService
 */
export type HealthService = {
    /**
     * Label
     */
    label: string;
    type: HealthServiceType;
    /**
     * Delta Days
     */
    delta_days: number;
    /**
     * Detail
     */
    detail: unknown;
};

/**
 * HealthServiceType
 */
export type HealthServiceType = 'PROCEDURE' | 'MEDICATION' | 'HEALTH_TOOL' | 'STOMATOLOGICAL_TOOL' | 'TRANSPORT' | 'HOSPITALIZATION' | 'SPA' | 'DEATH' | 'UNKNOWN';

/**
 * Patient
 */
export type Patient = {
    /**
     * Patient Id
     */
    patient_id: string;
    /**
     * Name
     */
    name: string;
};

/**
 * PatientFuture
 */
export type PatientFuture = {
    /**
     * Expected Health Services
     */
    expected_health_services: Array<HealthService>;
    /**
     * Probability
     */
    probability: number;
};

/**
 * PatientHistory
 */
export type PatientHistory = {
    /**
     * Received Health Services
     */
    received_health_services: Array<HealthService>;
};

/**
 * SuggestResult
 */
export type SuggestResult = {
    /**
     * Label
     */
    label: string;
    type: SuggestResultType;
};

/**
 * SuggestResultType
 */
export type SuggestResultType = 'PATIENT' | 'SERVICE_PROVIDER' | 'DRG';

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type SuggestSuggestGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/suggest';
};

export type SuggestSuggestGetResponses = {
    /**
     * Response Suggest Suggest Get
     *
     * Successful Response
     */
    200: Array<SuggestResult>;
};

export type SuggestSuggestGetResponse = SuggestSuggestGetResponses[keyof SuggestSuggestGetResponses];

export type GetPatientsPatientsGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/patients';
};

export type GetPatientsPatientsGetResponses = {
    /**
     * Response Get Patients Patients Get
     *
     * Successful Response
     */
    200: Array<Patient>;
};

export type GetPatientsPatientsGetResponse = GetPatientsPatientsGetResponses[keyof GetPatientsPatientsGetResponses];

export type GetPatientPatientsPatientIdGetData = {
    body?: never;
    path: {
        /**
         * Patient Id
         */
        patient_id: unknown;
    };
    query?: never;
    url: '/patients/{patient_id}';
};

export type GetPatientPatientsPatientIdGetErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetPatientPatientsPatientIdGetError = GetPatientPatientsPatientIdGetErrors[keyof GetPatientPatientsPatientIdGetErrors];

export type GetPatientPatientsPatientIdGetResponses = {
    /**
     * Successful Response
     */
    200: Patient;
};

export type GetPatientPatientsPatientIdGetResponse = GetPatientPatientsPatientIdGetResponses[keyof GetPatientPatientsPatientIdGetResponses];

export type GetPatientHistoryPatientsPatientIdHistoryGetData = {
    body?: never;
    path: {
        /**
         * Patient Id
         */
        patient_id: unknown;
    };
    query?: never;
    url: '/patients/{patient_id}/history';
};

export type GetPatientHistoryPatientsPatientIdHistoryGetErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetPatientHistoryPatientsPatientIdHistoryGetError = GetPatientHistoryPatientsPatientIdHistoryGetErrors[keyof GetPatientHistoryPatientsPatientIdHistoryGetErrors];

export type GetPatientHistoryPatientsPatientIdHistoryGetResponses = {
    /**
     * Successful Response
     */
    200: PatientHistory;
};

export type GetPatientHistoryPatientsPatientIdHistoryGetResponse = GetPatientHistoryPatientsPatientIdHistoryGetResponses[keyof GetPatientHistoryPatientsPatientIdHistoryGetResponses];

export type GetPatientFuturesPatientsPatientIdFuturesGetData = {
    body?: never;
    path: {
        /**
         * Patient Id
         */
        patient_id: string;
    };
    query?: {
        /**
         * Snapshot Events
         */
        snapshot_events?: number;
        /**
         * Top K
         */
        top_k?: number;
    };
    url: '/patients/{patient_id}/futures';
};

export type GetPatientFuturesPatientsPatientIdFuturesGetErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetPatientFuturesPatientsPatientIdFuturesGetError = GetPatientFuturesPatientsPatientIdFuturesGetErrors[keyof GetPatientFuturesPatientsPatientIdFuturesGetErrors];

export type GetPatientFuturesPatientsPatientIdFuturesGetResponses = {
    /**
     * Response Get Patient Futures Patients  Patient Id  Futures Get
     *
     * Successful Response
     */
    200: Array<PatientFuture>;
};

export type GetPatientFuturesPatientsPatientIdFuturesGetResponse = GetPatientFuturesPatientsPatientIdFuturesGetResponses[keyof GetPatientFuturesPatientsPatientIdFuturesGetResponses];

export type GetPatientEwsPatientsPatientIdEwsGetData = {
    body?: never;
    path: {
        /**
         * Patient Id
         */
        patient_id: unknown;
    };
    query?: never;
    url: '/patients/{patient_id}/ews';
};

export type GetPatientEwsPatientsPatientIdEwsGetErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type GetPatientEwsPatientsPatientIdEwsGetError = GetPatientEwsPatientsPatientIdEwsGetErrors[keyof GetPatientEwsPatientsPatientIdEwsGetErrors];

export type GetPatientEwsPatientsPatientIdEwsGetResponses = {
    /**
     * Response Get Patient Ews Patients  Patient Id  Ews Get
     *
     * Successful Response
     */
    200: Array<Ews>;
};

export type GetPatientEwsPatientsPatientIdEwsGetResponse = GetPatientEwsPatientsPatientIdEwsGetResponses[keyof GetPatientEwsPatientsPatientIdEwsGetResponses];